build: create SideBar component from source in Home.tsx
